+++ addons/a11y/src/components/Report/Item.js
@@ -8,19 +8,28 @@ import Info from './Info';
 import Tags from './Tags';
 import Elements from './Elements';
 
-const Wrapper = styled('div')({
+const Wrapper = styled('div')(({ theme }) => ({
   padding: '0 14px',
   cursor: 'pointer',
-  borderBottom: '1px solid rgb(234, 234, 234)',
-});
+  borderBottom: theme.mainBorder,
+}));
 
-const HeaderBar = styled('button')({
+const HeaderBar = styled('button')(({ theme }) => ({
   padding: '12px 0px',
   display: 'block',
   width: '100%',
   border: 0,
   background: 'none',
-});
+  color: 'inherit',
+
+  borderTop: '3px solid transparent',
+  borderBottom: '3px solid transparent',
+
+  '&:focus': {
+    outline: '0 none',
+    borderBottom: `3px solid ${theme.highlightColor}`,
+  },
+}));
 
 class Item extends Component {
   static propTypes = {