+++ addons/knobs/src/components/types/Number.js
@@ -2,7 +2,8 @@ import PropTypes from 'prop-types';
 import React from 'react';
 
 import styled from 'react-emotion';
-import debounce from 'lodash.debounce';
+
+import { Input } from '@storybook/components';
 
 const base = {
   boxSizing: 'border-box',
@@ -15,11 +16,6 @@ const base = {
   color: '#444',
 };
 
-const TextInput = styled('input')(base, {
-  display: 'table-cell',
-  width: '100%',
-  verticalAlign: 'middle',
-});
 const RangeInput = styled('input')(base, {
   display: 'table-cell',
   flexGrow: 1,
@@ -37,25 +33,12 @@ const RangeWrapper = styled('div')({
 });
 
 class NumberType extends React.Component {
-  constructor(props) {
-    super(props);
-    let { value } = props.knob;
-    if (value === null || value === undefined) {
-      value = '';
-    }
-    this.state = { value };
-
-    this.onChange = debounce(props.onChange, 400);
-  }
-
-  componentWillUnmount() {
-    this.onChange.cancel();
+  shouldComponentUpdate(nextProps) {
+    return nextProps.knob.value !== this.props.knob.value;
   }
 
   handleChange = event => {
     const { value } = event.target;
-
-    this.setState({ value });
 
     let parsedValue = Number(value);
 
@@ -63,52 +46,49 @@ class NumberType extends React.Component {
       parsedValue = null;
     }
 
-    this.onChange(parsedValue);
+    this.props.onChange(parsedValue);
   };
 
   render() {
     const { knob } = this.props;
-    const { value } = this.state;
 
     return knob.range ? (
       <RangeWrapper>
         <RangeLabel>{knob.min}</RangeLabel>
         <RangeInput
-          id={knob.name}
-          value={value}
+          value={knob.value}
           type="range"
           min={knob.min}
           max={knob.max}
           step={knob.step}
           onChange={this.handleChange}
         />
-        <RangeLabel>{`${value} / ${knob.max}`}</RangeLabel>
+        <RangeLabel>{`${knob.value} / ${knob.max}`}</RangeLabel>
       </RangeWrapper>
     ) : (
-      <TextInput
-        id={knob.name}
-        value={value}
+      <Input
+        value={knob.value}
         type="number"
         min={knob.min}
         max={knob.max}
         step={knob.step}
         onChange={this.handleChange}
+        size="flex"
       />
     );
   }
 }
 
-NumberType.defaultProps = {
-  knob: {},
-  onChange: value => value,
-};
-
 NumberType.propTypes = {
   knob: PropTypes.shape({
     name: PropTypes.string,
     value: PropTypes.number,
-  }),
-  onChange: PropTypes.func,
+    range: PropTypes.bool,
+    min: PropTypes.number,
+    max: PropTypes.number,
+    step: PropTypes.number,
+  }).isRequired,
+  onChange: PropTypes.func.isRequired,
 };
 
 NumberType.serialize = value => (value === null || value === undefined ? '' : String(value));