+++ app/react-native/src/server/config/webpack.config.prod.js
@@ -1,5 +1,6 @@
 import path from 'path';
 import webpack from 'webpack';
+import { getEnvironment } from 'universal-dotenv';
 import Dotenv from 'dotenv-webpack';
 import HtmlWebpackPlugin from 'html-webpack-plugin';
 import { indexHtmlPath } from '@storybook/core/server';
@@ -37,6 +38,7 @@ const getConfig = options => {
         storybookOptions: JSON.stringify(options),
       }),
       new webpack.optimize.DedupePlugin(),
+      new webpack.DefinePlugin(getEnvironment().webpack),
       new Dotenv({ silent: true }),
     ],
     module: {