+++ examples/react-native-typescript/src/__tests__/App.tsx
@@ -6,7 +6,5 @@ import App from '../App';
 import renderer from 'react-test-renderer';
 
 it('renders correctly', () => {
-  const tree = renderer.create(
-    <App />
-  );
+  renderer.create(<App />);
 });